% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_contact.R
\name{clean_contact}
\alias{clean_contact}
\title{Dispose of Uninformative Emails and Phones}
\usage{
clean_contact(df, email = NULL, email_exc = "abc@example.com",
  phone = NULL, phone_exc = "___-____")
}
\arguments{
\item{df}{Input dataframe to be cleaned of uninformative contacts.}

\item{email}{Name of the email address field.
Defaults to NULL.}

\item{email_exc}{Emails that are to be cleaned.
Defaults to a single vector of abc at example.com}

\item{phone}{Name of the phone number field.
Defaults to NULL.}

\item{phone_exc}{Phone numbers that are to be cleaned.
Defaults to "___-____".}
}
\value{
Output dataframe with cleaned contacts.
}
\description{
This function takes a dataframe that contains personal contacts such as
emails and addresses, and cleans out uninformative contacts. For example,
if a particular email address is input as fake or a default value from
some particular setting, these values are replaced as NA.
Similarly for phone numbers, if they are likely fake such as (999)999-9999
or contains only area numbers, they are replaced as NA.
}
