% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_snapshot.R
\name{clean_snapshot}
\alias{clean_snapshot}
\title{Clean Snapshots and Store Them}
\usage{
clean_snapshot(date_df = NULL, start = "2018-04-26",
  end = "2021-01-01", path = "7z",
  pattern = "^(?=.*Cntywd_)(?!.*Hist)", file_type = ".txt",
  path_clean = "clean_df", clean_prefix = "df_cleaned_",
  clean_suffix = "", save_type = c("rda", "fst"),
  format = "\%m\%d\%y", rec = FALSE, period = 1,
  file_prefix = "Cntywd_", varnames = NULL, date = NULL,
  date_order = "mdy", num = NULL, first = "szNameFirst",
  voter_prefix = "sVoterTitle", gender = "sGender",
  email = "szEmailAddress", email_exc = c("abc@example.com"),
  phone = "szPhone", phone_exc = "___-____", ...)
}
\arguments{
\item{date_df}{List of snapshots. Defaults to NULL,
in which case the function will detect all snapshots available.}

\item{start}{The start date of the first snapshot.
Defaults to April 26, 2018.}

\item{end}{The end date of the last snapshot.
Defaults to Jan 1, 2021.}

\item{path}{Path where all snapshots are stored.
Defaults to subfolder 7z.}

\item{pattern}{Regular expression of the file pattern to find.
Defaults to a particular pattern of OCROV files.}

\item{file_type}{File type.
Defaults to .txt.}

\item{path_clean}{Path where cleaned snapshots would be stored.
Defaults to "clean_df".}

\item{clean_prefix}{File prefixes for cleaned snapshots.
This replaces the existing file prefix. Defaults to "df_cleaned_".}

\item{clean_suffix}{File suffixes for cleaned snapshots.
Defaults to empty string.}

\item{save_type}{How to export the cleaned dataframe.
Defaults to Rda and fst.}

\item{format}{Format of the date in the snapshot file names.
Defaults to "\%m\%d\%y".}

\item{period}{Period/interval between each snapshot---
whether daily, weekly, and so on.
Defaults to 1 (equivalent to "day"). Any valid input for base seq.Date
by argument is allowed.}

\item{file_prefix}{File name prefix.
Defaults to Cntywd_.}

\item{varnames}{All variables to be cleaned.
Defaults to NULL.}

\item{date}{Date variables.
Defaults to NULL.}

\item{date_order}{Order of the date variable, if string format.}

\item{num}{Numeric variables.
Defaults to NULL.}

\item{first}{Variable containing first names.
Defaults to "szNameFirst".}

\item{voter_prefix}{Variable containing self-reported personal prefixes.
Defaults to "sVoterTitle".}

\item{gender}{Variable containing original gender entry.
Defaults to "sGender".}

\item{email}{Name of the email address field.
Defaults to "szEmailAddress".}

\item{email_exc}{Emails that are to be cleaned.
Defaults to a single vector of abc at example.com}

\item{phone}{Name of the phone number field.
Defaults to "szPhone".}

\item{phone_exc}{Phone numbers that are to be cleaned.
Defaults to "___-____".}

\item{...}{Other arguments to be passed to snapshot_import.}

\item{recursive}{Whether to find files recursively.
Defaults to FALSE.}
}
\value{
Output dataframe with cleaned contacts.
}
\description{
This function takes requested snapshots, import them, clean them, and
export them into Rda and/or fst objects for future calls by vrmatch function.
}
