% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_match.R
\name{exact_match}
\alias{exact_match}
\title{Extract Exact Matches Between Two Snapshots}
\usage{
exact_match(df_list, dfA = "dfA", dfB = "dfB")
}
\arguments{
\item{df_list}{A named list that contains two consecutive snapshots.}

\item{dfA}{Name of the first snapshot in df_list for exact matching.
Defaults to "dfA".}

\item{dfB}{Name of the second snapshot in df_list for exact matching.
Defaults to "dfB".}
}
\value{
A list of three dataframes: exact_match, mismatch_A, and mismatch_B.
}
\description{
This function takes a list of two snapshots, performs exact matching
by dplyr::inner_join, and outputsa list of three dataframes:
the first dataframe contains records that are exact matches between the
two dataframes. The second dataframe contains records that are only in the
first snapshot. The third dataframe contains records that are only in the
second snapshot. This is to take the load off record linkage when the
snapshots are granular, such as daily snapshots.
}
