% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_match.R
\name{id_match}
\alias{id_match}
\title{Extract Exact Matches by ID Fields Between Two Dataframes}
\usage{
id_match(df_list, ids = "lVoterUniqueID", dfA = "mismatch_A",
  dfB = "mismatch_B")
}
\arguments{
\item{df_list}{A named list that contains two dataframes
to check for ID matches.}

\item{ids}{A vector of internal IDs to iterate for exact match extracts.
Defaults to "lVoterUniqueID".}

\item{dfA}{Name of the first snapshot in df_list for exact matching in IDs.
Defaults to "mismatch_A".}

\item{dfB}{Name of the second snapshot in df_list for exact matching in IDs.
Defaults to "mismatch_B".}
}
\value{
the data frame with IDs corrected for
}
\description{
This function takes a named list of two dataframes, performs exact matching
by specified ID field(s), and outputs a list of four dataframes:
the original dataframes are split into ID matches and non-matches.
The original dataframes would be rid of records that match in IDs.
The additional dataframes of "id_match_A" and "id_match_B" would contain
records with matches in ID in the respective input dataframes.
}
