% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_colclasses.R
\docType{data}
\name{oc_colclasses}
\alias{oc_colclasses}
\title{Column Classes for OCROV Data}
\format{A list of length 62.}
\usage{
oc_colclasses
}
\description{
A dataset containing the list of column classes for data from
Orange County Registrar of Voters.
}
\keyword{datasets}
