% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Assessing the Performance of Record Linkage with a True Match Variable}
\usage{
performance(date_df, path_matches = "matches",
  date_label = "date_label", ids = c("lVoterUniqueID", "sAffNumber"),
  cond = "or", inter_id = "inter_id")
}
\arguments{
\item{date_df}{Dataframe of list of snapshots.}

\item{path_matches}{Path where the match outcomes are output to.
Defaults to "matches".}

\item{date_label}{Labels for dates (i.e., snapshot IDs), in `date_df`.
Defaults to "date_label".}

\item{ids}{A vector of true match fields.
Defaults to c("lVoterUniqueID", "sAffNumber").}

\item{cond}{Whether the supplied true match fields should be used with a
OR condition or AND. Defaults to "or".}

\item{inter_id}{Interm ID for dplyr::bind_row for purrr::output.
Defaults to "inter_id".}
}
\value{
A summary dataframe of performance assessment of all user-specified
snapshot matches.
}
\description{
This function takes the match snapshot(s) that are outputs from vrmatch,
and using a variable that shows the true match status, i.e., for instance
an internally validated voter ID field, assesses the performance of the
match. We can use multiple true match variables---for instance,
a combination of the internal voter ID variable (field A) and
the registration affidavit number variable (field B).
The records will be true matches if either A or B matches. Vice versa,
we can also garner true matches based on all true match variables being
matched.
}
\details{
Note that this differs from fastLink::summary in that we are utilizing
an externally supplied true match variable.
}
