% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_list.R
\name{snapshot_list}
\alias{snapshot_list}
\title{List the Available Snapshots}
\usage{
snapshot_list(start = "2018-04-26", end = "2021-01-01", path = "7z",
  pattern = "^(?=.*Cntywd_)(?!.*Hist)", file_type = ".txt",
  format = "\%m\%d\%y", recursive = FALSE, period = 1,
  file_prefix = "Cntywd_")
}
\arguments{
\item{start}{The start date of the first snapshot.
Defaults to April 26, 2018.}

\item{end}{The end date of the last snapshot.
Defaults to Jan 1, 2021.}

\item{path}{Path where all snapshots are stored.
Defaults to subfolder 7z.}

\item{pattern}{Regular expression of the file pattern to find.
Defaults to a particular pattern of OCROV files.}

\item{file_type}{File type.
Defaults to .txt.}

\item{format}{Format of the date in the snapshot file names.
Defaults to "%m%d%y".}

\item{recursive}{Whether to find files recursively.
Defaults to FALSE.}

\item{period}{Period/interval between each snapshot---
whether daily, weekly, and so on.
Defaults to 1 (equivalent to "day"). Any valid input for base seq.Date
by argument is allowed.}

\item{file_prefix}{File name file_prefix.
Defaults to Cntywd_.}
}
\value{
A dataframe that contains available snapshots.
}
\description{
This function lists currently available snapshots, assuming a common
file pattern and snapshot IDs by date. The current default values are
tailored to the Orange County setup.
}
