% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrmatch.R
\name{vrmatch}
\alias{vrmatch}
\title{Voter Registration Database Snapshot Matching}
\usage{
vrmatch(date_df, exact_exclude = TRUE, sample = FALSE, block = FALSE,
  path_clean = "clean_df", path_changes = "changes",
  path_reports = "reports", path_matches = "matches",
  clean_prefix = "df_cleaned_", clean_suffix = "",
  exist_files = FALSE, varnames, varnames_str, varnames_num = NULL,
  varnames_id = NULL, partial.match = NULL, varnames_block = NULL,
  vars_change = NULL, n.cores = NULL, file_type = ".Rda",
  date_label = "date_label", nrow = "nrow", seed = 123, ...)
}
\arguments{
\item{date_df}{Dataframe of list of snapshots.}

\item{exact_exclude}{Whether to exclude full exact matches between snapshots
when doing probabilistic record linkage. Defaults to TRUE.}

\item{sample}{Whether to add random samples of full exact matches or ID
matches to correct for underlying population's value distributions for
each field. Defaults to FALSE.}

\item{block}{Whether to employ blocking.
Defaults to FALSE.}

\item{path_clean}{Path to the cleaned snapshots.
Defaults to "clean_df".}

\item{path_changes}{Path where the extracted changes are output to.
Defaults to "changes".}

\item{path_reports}{Path where the summarized changes are output to.
Defaults to "reports".}

\item{path_matches}{Path where the match outcomes are output to.
Defaults to "matches".}

\item{clean_prefix}{File prefixes for cleaned snapshots.
Defaults to "df_cleaned_".}

\item{clean_suffix}{File suffixes for cleaned snapshots.
Defaults to empty string.}

\item{exist_files}{Whether previously performed match outcomes exist.
Defaults to FALSE.}

\item{varnames}{Variables to perform probabilistic record linkage.}

\item{varnames_str}{String variables for matching.}

\item{varnames_num}{Numeric variables for matching.
Defaults to NULL, in which case it will be setdiff(varnames, varnames_str).}

\item{varnames_id}{Voter IDs variables, if any exists, and is to be excluded
from PRL when IDs match.}

\item{partial.match}{Variables to be partially matched.
Defaults to all varnames_str.}

\item{varnames_block}{Nested list of variables or their combinations
for blocking passes.}

\item{vars_change}{Variables to track changes of.
Defaults to NULL, which will then track all variables.}

\item{n.cores}{Number of cores to parallelize the matching.
Defaults to half the existing threads.}

\item{file_type}{Input file types.
Defulats to .Rda.}

\item{date_label}{Labels for dates (i.e., snapshot IDs), in `date_df`.
Defaults to "date_label".}

\item{nrow}{Name of list element which will contain the number of rows
of the input list dataframes.}

\item{seed}{Seed to set. Defaults to 123.}

\item{...}{Other parameters for fastLink.}
}
\value{
A nested list of matched dataframes, fastLink output, and arguments.
}
\description{
This function performs probabilistic record linkage between all user-supplied
consecutive snapshots of the voter file. Note that the default option is to
exclude exact matches of all fields between two snapshots when performing
the record linkage, for computational reasons.
}
