% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes_report.R
\name{changes_report}
\alias{changes_report}
\title{Report Changes Between Snapshots}
\usage{
changes_report(changes, varnames = NULL, nrow = "nrow")
}
\arguments{
\item{changes}{Named list outputed from change_extract.}

\item{varnames}{Variables to track changes.
Defaults to NULL.}

\item{nrow}{Name of list element which will contain the number of rows
of the input list dataframes.}
}
\value{
Dataframe with changes' summary statistics.
}
\description{
This function takes the extracted matches from changes_extract,
and then summarizes what has changed between the snapshots compared.
}
