% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_import.R
\name{snapshot_import}
\alias{snapshot_import}
\title{Importing Snapshots into List/Dataframe}
\usage{
snapshot_import(path = ".", prefix = "", units, suffix = "",
  file_type = ".txt", col_classes = NULL, n_max = Inf,
  enc = "ISO-8859-1", del = "\\t", quote = "", col_names = TRUE,
  na = "", ref = NULL, ...)
}
\arguments{
\item{path}{File path to target.
Defaults to current directory.}

\item{prefix}{File name prefix.}

\item{units}{Snapshot ID e.g. 20180426.
Defaults to empty string.}

\item{suffix}{File name suffix.
Defaults to empty string.}

\item{file_type}{Input file type. Currently takes in .txt and .csv.
Defaults to .txt.}

\item{col_classes}{A list of column classes for faster import, if available.
Defaults to NULL.}

\item{n_max}{Maximum number of records to import.
Defaults to Inf.}

\item{enc}{Encoding. Default is ISO-8859-1, to account for certain names.}

\item{del}{Delimiter. Default is tab.}

\item{quote}{Set of quoting characters. Default is an empty string.}

\item{col_names}{Whether column names exist in the dataframe.
Defaults to TRUE.}

\item{na}{NA characters to be interpreted.
Defaults to empty string.}

\item{ref}{Reference for snapshot imports when more than two are imported.
Defaults to English lowercase alphabets.}

\item{...}{Other arguments to be passed to
readr::read:csv or readr::read_delim}
}
\value{
A list of dataframes of the desired data. If only a single unit is
input, a single dataframe.
}
\description{
This function imports consecutive snapshots of the data into a list output.
If there is only one snapshot input requested, it will return a single
dataframe as is. Unless otherwise supplied, the output list will reference
the snapshots as dfA, dfB, dfC, and so on, to ensure that the output can be
supplied to other functions generically.
}
